backend/pyproject.toml
[project]
name = "aise2"
version = "0.1.0"
description = "AISE 2.0 Backend"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "deepagents>=0.4.12",
    "fastapi[standard]>=0.135.2",
    "loguru>=0.7.3",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "openai>=2.30.0",
    "alembic>=1.18.4",
    "psycopg2-binary>=2.9.11",
    "pgvector>=0.4.0",
    "minio>=7.2.0",
    "pymupdf>=1.25.0",
    "python-docx>=1.1.0",
    "python-pptx>=1.0.0",
    "openpyxl>=3.1.0",
    "tiktoken>=0.9.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]

